+++ app/src/renderer/components/wallet/PageSend.vue
@@ -30,9 +30,7 @@ page(title='Send')
             v-model='fields.address'
             placeholder='Address')
         form-msg(name='Address' type='required' v-if='!$v.fields.address.required')
-        form-msg(name='Address' type='exactLength' length='40'
-          v-if='!$v.fields.address.minLength || !$v.fields.address.maxLength')
-        form-msg(name='Address' type='alphaNum' v-if='!$v.fields.address.alphaNum')
+        form-msg(name='Address' type='bech32' :body="bech32error" v-else-if='!$v.fields.address.bech32Validate')
 
       form-group(:error='$v.fields.amount.$error'
         field-id='send-amount' field-label='Amount')
@@ -52,13 +50,8 @@ page(title='Send')
 </template>
 
 <script>
-import {
-  required,
-  between,
-  minLength,
-  maxLength,
-  alphaNum
-} from "vuelidate/lib/validators"
+import b32 from "scripts/b32"
+import { required, between } from "vuelidate/lib/validators"
 import { mapActions, mapGetters } from "vuex"
 import Btn from "@nylira/vue-button"
 import Field from "@nylira/vue-field"
@@ -100,6 +93,7 @@ export default {
     }
   },
   data: () => ({
+    bech32error: null,
     fields: {
       address: "",
       amount: null,
@@ -126,13 +120,14 @@ export default {
       let zoneId = this.fields.zoneId
       try {
         // if address has a slash, it is IBC address format
-        let type
-        if (this.lastHeader.chain_id !== zoneId) {
-          type = "ibcSend"
-          address = `${zoneId}/${address}`
-        } else {
-          type = "send"
-        }
+        let type = "send"
+        // TODO reenable when we have IBC
+        // if (this.lastHeader.chain_id !== zoneId) {
+        //   type = "ibcSend"
+        //   address = `${zoneId}/${address}`
+        // } else {
+        //   type = "send"
+        // }
         await this.sendTx({
           type,
           to: address,
@@ -155,6 +150,16 @@ export default {
         })
       }
     },
+    bech32Validate(param) {
+      try {
+        b32.decode(param)
+        this.bech32error = null
+        return true
+      } catch (error) {
+        this.bech32error = error.message
+        return false
+      }
+    },
     ...mapActions(["sendTx"])
   },
   props: ["denom"],
@@ -169,9 +174,7 @@ export default {
       fields: {
         address: {
           required,
-          minLength: minLength(40),
-          maxLength: maxLength(40),
-          alphaNum: alphaNum
+          bech32Validate: this.bech32Validate
         },
         amount: {
           required,
@@ -183,11 +186,11 @@ export default {
     }
   },
   watch: {
-    // TODO should not be necessary?
-    // if the zoneId gets added at a later time
-    "wallet.zoneIds": () => {
-      this.fields.zoneId = this.wallet.zoneIds[0]
-    }
+    // TODO ignored while we don't have IBC
+    // // if the zoneId gets added at a later time
+    // "wallet.zoneIds": () => {
+    //   this.fields.zoneId = this.wallet.zoneIds[0]
+    // }
   }
 }
 </script>