+++ test/unit/specs/components/wallet/LiTransaction.spec.js
@@ -15,7 +15,7 @@ describe("LiTransaction", () => {
                 amount: 1234
               }
             ],
-            sender: "otherAddress"
+            sender: "tb1da6xsetjg9jxgun9wdesexv05j"
           }
         ],
         outputs: [
@@ -26,13 +26,13 @@ describe("LiTransaction", () => {
                 amount: 1234
               }
             ],
-            recipient: "myAddress"
+            recipient: "tb1d4u5zerywfjhxuc9nudvw"
           }
         ]
       },
       time: null
     },
-    address: "myAddress"
+    address: "tb1d4u5zerywfjhxuc9nudvw"
   }
 
   beforeEach(() => {
@@ -59,7 +59,7 @@ describe("LiTransaction", () => {
                   amount: 1234
                 }
               ],
-              sender: "myAddress"
+              sender: "tb1d4u5zerywfjhxuc9nudvw"
             }
           ],
           outputs: [
@@ -70,13 +70,13 @@ describe("LiTransaction", () => {
                   amount: 1234
                 }
               ],
-              recipient: "otherAddress"
+              recipient: "tb1da6xsetjg9jxgun9wdesexv05j"
             }
           ]
         },
         time: Date.now()
       },
-      address: "myAddress"
+      address: "tb1d4u5zerywfjhxuc9nudvw"
     })
     expect(wrapper.find(".ni-li-tx").classes()).toContain("ni-li-tx-sent")
   })
@@ -101,7 +101,7 @@ describe("LiTransaction", () => {
                   amount: 42
                 }
               ],
-              sender: "otherAddress"
+              sender: "tb1da6xsetjg9jxgun9wdesexv05j"
             }
           ],
           outputs: [
@@ -120,13 +120,13 @@ describe("LiTransaction", () => {
                   amount: 42
                 }
               ],
-              recipient: "myAddress"
+              recipient: "tb1d4u5zerywfjhxuc9nudvw"
             }
           ]
         },
         time: Date.now()
       },
-      address: "myAddress"
+      address: "tb1d4u5zerywfjhxuc9nudvw"
     })
     expect(wrapper.findAll(".tx-coin").length).toBe(3)
     expect(