+++ test/unit/specs/lcdClient.spec.js
@@ -60,10 +60,7 @@ describe("LCD Client", () => {
     axios.get = jest.fn().mockReturnValueOnce(
       Promise.reject({
         response: {
-          data: {
-            error: "foo",
-            code: 123
-          }
+          data: "foo"
         }
       })
     )
@@ -72,7 +69,6 @@ describe("LCD Client", () => {
       await await client.listKeys()
     } catch (err) {
       expect(err.message).toBe("foo")
-      expect(err.code).toBe(123)
     }
     expect(axios.get.mock.calls[0]).toEqual([
       "http://localhost:8998/keys",
@@ -93,10 +89,7 @@ describe("LCD Client", () => {
     axios.get = jest.fn().mockReturnValueOnce(
       Promise.reject({
         response: {
-          data: {
-            error: "account bytes are empty",
-            code: 1
-          }
+          data: "account bytes are empty"
         }
       })
     )
@@ -108,10 +101,7 @@ describe("LCD Client", () => {
     axios.get = jest.fn().mockReturnValueOnce(
       Promise.reject({
         response: {
-          data: {
-            error: "something failed",
-            code: 1
-          }
+          data: "something failed"
         }
       })
     )